add attributes and functions triangle
