--- HW1/oop.py
@@ -1,15 +1,0 @@
-"""Создать базовый класс геометрической фигуры (Figure).
-Реализовать на его основе классы геометрических фигур:
-Треугольник, Прямоугольник, Квадрат, Круг (Triangle, Rectangle, Square, Circle)."""
-
-
-# Создаем класс
-class Figure:
-    pass
-
-
-# Создаем объекты
-Triangle = Figure()
-Rectangle = Figure()
-Square = Figure()
-Circle = Figure()
