+++ home_work_1/test_hw1.py
@@ -1,0 +1,2 @@
+import pytest
+
